+++ src/kernel/DEMKinematicMisc.cu
@@ -66,7 +66,7 @@ DEME_KERNEL void computeMarginFromAbsv_implTri(deme::DEMSimParams* simParams,
                                                const float* absAngVel_owner,
                                                float* ts,
                                                unsigned int* maxDrift,
-                                               double* maxTriTriPenetration,
+                                               float* maxTriTriPenetration,
                                                bool meshUniversalContact,
                                                size_t n) {
     size_t triID = blockIdx.x * blockDim.x + threadIdx.x;
@@ -83,7 +83,7 @@ DEME_KERNEL void computeMarginFromAbsv_implTri(deme::DEMSimParams* simParams,
         // Compute additional margin based on per-triangle max tri-tri penetration if meshUniversalContact is enabled.
         // This is needed as our meshed particle representation is surface only, so we need to account for existing
         // penetration length in our future-proof contact detection, always.
-        double penetrationMargin = maxTriTriPenetration[triID];
+        double penetrationMargin = (double)maxTriTriPenetration[triID];
         penetrationMargin = (meshUniversalContact && penetrationMargin > 0.0) ? penetrationMargin : 0.0;
         // Clamp penetration margin to the maximum allowed value to prevent super large margins
         if (penetrationMargin > simParams->capTriTriPenetration) {
